add Nim versions of many C examples
These examples try to stay as close to the C examples as possible to
serve as a reference. It shows that the low level API to the C
functions is exactly the same, including all C macro definitions.

Higher level abstractions, to reach the simplicity of the h5py
examples, will be done separately (after it is implemented...).