add option to create dataset for 1D col, using int instead of tuple
Adds the ability to create a dataset for a single column of data, by
not having to hand a tuple of (<value>, 1). Now we can either hand a
tuple describing the shape of the data to be written or simply hand an
integer for the length of the 1D data.