add basic option to write arraymancer.Tensor to HDF5 file via `[]=`
Adds basic functionality to assign an arraymancer tensor to a
H5DataSet. Internally, it checks the dimensions of the tensor,
squeezes it and accesses the raw data using
arraymancer.Tensor.toRawSeq.

Be careful not to hand a view of a tensor, as the tensor.shape
function acts on the original tensor, not on the view, as does the
toRawSeq function!