add support to write chunked storage, which can be resized
Adds support to create not only contiuous memory datasets, but also
chunked datasets (if desired with a maximum size), which have the
advantage of being resizable. Includes a `resize` proc to do just
that.

In addition starts including custom exceptions to deal with failed
calls to the HDF5 library etc.