fix bug causing hyperslab writes to end up at wrong location
A bug caused data written by the write_hyperslab function to end up in
the first fitting array (of the shape of the hyperslab to write in the
dataset), instead of the `offset` location desired. This was due to
- the wrong dataspace used for selection
- the wrong shape for the memory space used
the combination of both resulted in a 'correct' selection of a
hyperslab, so that the HDF5 library did not complain, but ended up
writing only to the first elements, since the data to be written was
less than the /wrongly/ selected hyperslab amounted to.