introduce hotfix to create hardlink of object, not currently active
Introduces a short hotfix, which allows the creation of hardlinks
of paths, which exist in file, but of which the H5FileObj is not aware
atm. This by itself is fine, since we check whether the group exists
before 'creating' it (i.e. making H5FileObj aware of it). But the way
it's implemented atm renders the getObjectIdByName function useless
for cases, in which it is called for a group which in fact does not
exist, because it will simply be created instead (and an ID returned).

FIX THIS