add compiler flag -d:H5_LEGACY to keep support for 1.8
Adds the optional compiler flag -d:H5_LEGACY in order to support both
HDF5 1.8 as well as the HDF5 1.10 release. By default the library
builts support for HDF5 1.10. If an older shared library is installed,
build with the flag active. The wrong setting will most likely result
in a
`could not import: <HDF5 var name>`
error.