demand dset datatype on read to be exact
Set the required datatype to match when reading full datasets back to
specifically failing, if the given type (as string) does not exactly
match the type inferred from the H5 type.

Until a better solution (taking care about potential ambiguous
datatype names) is in place, this seems the sanest choice.