modify group get s.t. it also checks file for existence of group
Previously we only checked whether the H5FileObj was already aware of
the group we're trying to access.

Now we check the file, in case the group table does not contain the
group, we're trying to get, and in case it exists in file we create
the group for the H5FileObj group table.