add iterator for datasets in groups, requires H5FileObj reference
Adds an iterator to iterate over datasets in a group. This currently
only works for datasets in the group itself, not recursively for
subgroups.

In order to do this such that it works without having the user call
`visit_file` beforehand, but also not giving the H5FileObj to the
iterator as well, a reference to the group's H5FileObj parent is
added.

Currently comes with a few design issues, as explained in Notes.org.

Also necessary to include forward declaration for visit_file to avoid
import problems, due to cyclic imports.