add proc to read hyperslab from dataset using H5 notation (issue #4)
Adds a proc to read an arbitrary hyperslab from a dataset. It
currently uses the normal HDF5 notation for hyperslabs (see HDF5
documentation sec. 7.4.1.1). A convenience notation using `[]` will be
added in the future.

The proc defaults to `full_output = false` behaviour, which means the
hyperslab is returned as a 1D sequence /only/ containing the read
elements.

Using the `full_output = true` option, a 1D sequence of the full size
of the datast is returned, which is zero everywhere except at the
locations specified by the hyperslab selection.