add `contains` proc for H5FileObj and H5Group
`contains` proc allows to check for existence of a group or a dataset
in a H5 file or a group.

This allows the usage of `in` and `notin` convenience keywords on a
file or group object.

For a group both a full or relative path is supported, however lookups
deeper than 1 level are not supported for groups (call on H5FileObj
for that).