add convenience proc `[]` to read scalar element from dataset
Adds a convenience proc to read a scalar element from a H5DataSet
given a single index value, using `[]` notation. Note that for N
dimensional datasets, the proc tries to always use the same value
`ind` for each dimension. If a dimension is too small, the last
element is taken instead.