change H5Attributes Table to ref Table
This change is done to fix problems, which arise if a H5 file was
visited (i.e. there is a group and dataset ref table) and we then
iterate over groups to add / update attributes. In case we end up
accessing the same group twice (without keeping the variable around),
we previously ended up with the case in which the just written
attribute was not known to the H5Attributes object stored in the ref
group of H5FileObj.

This way, a change to the ref Table (of the local group object)
immediately affects the ref Table of the ref Group in the H5FileObj.