change `getParentId` from template to proc, clean up
Changes the `getParentId` template to a procedure and cleans up its
code. Previously unecessary checks might have introduced bugs, due to
ignoring the case in which it is called for a `H5DataSet` instead of
raising an error. Unncessary because the code used for `H5Group` can
be used for `H5DataSet` as well.