rearange `create_dataset` to get dtype str rep from H5 dtype (cf #9)
Instead of using the given Nim `typedesc` and convert it to a string
representation, we now first create the dataset in the file and then
use its dataset id, get the datatype id and use that to get a string
representation, in the same way we do it when reading a dataset from a
file. This gets rid of (some) mismatches between datatypes when
writing them and using the just created H5DataSet objects vs. reading
a H5DataSet from a file and using that (issue #9).