add (hacky) proc to copy attributes from one obj to another
This adds a quite ugly way to copy attributes of one object to
another. While in principle simple, in practice we need to check for
the type of each attribute before we can read the value to write
it. To accomplish that, we introduce a `withAttr` template, which
simply performs a case on the `AnyKind` field of each attribute and
reads the data, injects an `attr` variable into the calling scope and
we use that to write the value.

TODO: this way is ugly, simply because we can rely on the native
H5Ocopy procedure instead. Implement and exchange by that!