add `convertType` proc to get a converter proc for numeric dsets
Given a numerical dataset (no char or string), `convertType` returns a
proc, which can be used to convert the data in the handed dataset to
an arbitrary other type. The conversion is done in Nim using `mapIt`
and applying the target datatype by normal conversion of
`it.targetType`.
Note that for the case where the actual datatype in the dataset is the
same as the target datatype, the resulting converter proc is a no-op.

This proc is useful to get a specific datatype, even if the actual
type in the file is not yet known and without having to resort to a
manual case on e.g. `dtypeAnyKind`.