add `dataspace_id` proc to get space id of a given dataset
Instead of keeping track of the dataspace id of a given dataset in the
H5DataSet object, we instead define a proc to get the correct
dataspace for a dataset using the `H5Dget_space` C procedure. This
allows us a cleaner interface without the need to update the dataspace
id all the time (and potentially keeping a stale space id around). It
also allows for several more removals of `var` requirements for datasets.