if no data given to `write_hyperslab` simply return, add test
If no data is given in write hyperslab, simply return instead of
causing a HDF5LibraryError when trying to create an empty simple
dataspace (and a following out of bounds error accessing `mdata[0]` if
`mdata.len == 0`).

Better: Parse the hyperslab selection and make sure the hyperslab
covers exactly as many elements as given in the data sequence.