fix dataset creation property list not being read
There was a typo in the `get` procedure for `H5Datasets`, which caused
us to assign the dataset creation property list to the field for the
dataset access property list (in effect overriding that).

This caused e.g. a problem when trying to resize a chunked dataset,
in a freshly opened file (i.e. w/o having created said dataset on this opening).