assert that parent has a valid id in `initH5Attributes`
There's no point in creating a H5Attributes object, if the given
parent id is invalid. We have an assertion in place to make to signal
that written code should make sure that this is never called with an
invalid id. The caller is responsible for checking for existence/
creating the parent for which the attributes are for!