make all objects ref objects
This change was long overdue.

HDF5 objects track all sorts of internal state to keep track of the
objects within the HDF5 C libray. This means for a large number of
operations, which don't change the given object from a user
perspective a `var H5*` object still had to be available, which often
meant having to declare e.g. a group or dataset as `var` although
apparently nothing that changes the objet would take place,
e.g. reading a simple dataset.

With every object being a `ref object` we kind of circumvent this,
because the immutability of a `ref object` refers to the reference
itself and not the object the reference points to.

Also from a more abstract point of view this makes a whole lot more
sense. These objects are not objects stored in a H5 file, but are
just *references* to objects in a file. Thinking about immutability of
such objects when the HDF5 C library does all sorts of crazy stuff in
the backend is almost ridiculous.

Also the tables, which store the different subgroups and datasets are
now proper `TableRef` instead of `ref Table` fields.
Although the content of these still needs to be looked at again...