remove dependency on typetraits, typeinfo, add custom enum
Instead of using the `AnyKind` enum, which is a holed enum, define our
own. Also don't need to workaround using a tmp variable of a dtype and
then use `toAny.kind` to get the correct `AnyKind`. Just use custom
macro to get name of type (which can extract it from generic T) and
parse to new enum.