replace `csize` by `csize_t`
This is an important change. Previously in 2019:
https://github.com/nim-lang/Nim/issues/12187
the Nim documentation stated that `csize` was the same as `size_t` in
C. However, `csize` was always a signed integer.

I trusted this when wrapping the HDF5 library (it was my first Nim
project :P) which means any `csize` arguments that overflowed were
always buggy.
The only place where I stumbled on problems though are `H5T_VARIABLE`
which is defined as `csize_t.high` (or previously `csize.high`). This
then of course was different from the actual value expected for this
"special" value in the HDF5 library.

This commit should hopefully replace all occurences of `csize` by `csize_t`.