`visit_file` doesn't open dsets/groups anymore, `opened` field added
Datasets and Groups are now aware of being open. In particular opening
every single dataset and group in case the file was visited (in which
case we only care about knowing which datasets and groups actually
exist) now doesn't open everything anymore.

`close` procs for datasets and groups are added.