introduce different distinct types for different `hid_t`
This is the largest change in `nimhdf5` in a long time.
We introduce different distinct types for all the different
purposes (at least the one we use) for `hid_t` in the HDF5
library. This gives us a lot more type safety, allows to properly
write `close` procedures, helps with readability of intention, etc.

Also includes cleanups of `create_dataset` & `create_group`.

Finally, includes a rename of the `dset_raw` and `shape_raw` fields in `create_dataset`.