fixup `=destroy` hooks to clean up properly
We never destroyed the references to the tables, nor the string and
seq fields.

This might cause GC problems on refc and under problems under orc.

At least it (seemed) to have caused a regression when dealing with H5
files that for some reason something attempts to close already closed
objects in a file.