add custom isOpen check for groups, datasets, files
As it may be unsafe to depend on our `opened` field, it's better to
make sure the object really *is* opened in the file.

So we retrieve all open identifiers and compare with our object.

Still: we may want to think about changing the logic to the same idea
as we do for attributes already:
Never have the dataset or group open longer than necessary. Open for
each transaction and close again.