add SWMR support, turn file access consts into enum, extend ObjectKind enum
This adds support for SWMR via opening of the file.

This is achieved by a rather large change to the logic of object kinds
& file access constants:

Instead of just dealing with a bunch of constants, we introduce /
extend enums for `ObjectKind` and `AccessKind`, which are then stored
in a set.
This allows for better handling in the code as we don't have to
manually perform any bitwise `or` operations etc.

We need a custom macro to iterate over the holed enum though.