do not rely on `opened` field, too easy to disagree with H5 lib
Our opened has too many ways in which it might not be synchronized
with the H5 library. Therefore it is better to not rely on it at
all. In particular thanks to the `=destroy` hook we simply now close
any dataset / attribute etc anyway when it goes out of scope. Better
to purely check what the library says and act on that.