replace distinct `hid_t` types by traced 'fat' objects
The basic idea here is the following:
The `hid_t` identifiers all refer to objects that live in the H5
library (and possibly in a file). In our previous approach we kept
track of different types by using `distinct hid_t` types. That's great
because we cannot mix and match the wrong type of identifiers in a
given context.
However, there are real resources underlying each identifier. Most
identifiers require the user to call a `close` / `free` type of
routine. While we can associate a destructor with a `=destroy` hook to
a `distinct hid_t` (with `hid_t` just being an integer type), the
issue is *when* that destructor is being called. In this old way the
identifier is a pure value type. If an identifier is copied and the
copy goes out of scope early, we release the resource despite still
needing it!
Therefore, we now have a 'fat' object that knows its internal
id (just a real `hid_t`) and which closing function to call. Our
actual IDs then are `ref objects` of these fat objects.
That way we get sane releasing of resources in the correct moments,
i.e. when the last reference to an identifier goes out of scope. This
is the correct thing to do in 99% of the cases.

Note: this commit is standalone and only defines the types. It does
not yet implement its usage. That's the next commits to keep this one separate.