refactor `write`, `read` procs into one interface, improve nested objs
Write and read now have essentially a single proc to interface with
the H5 library. Cleans up the code quite a bit and makes it more
understandable.

More importantly though, this majorly improves the support for nested
objects to be stored in an H5 file.
Types with string fields or nested sequences are now supported. These
will be copied over to a suitable, equivalent (in terms of H5 types)
tuple type. The same is done for reading types.

This is why `type_utils` and `copyflat` were added.