fix very hard to detect segfault on Windows
Wow, this was hard to understand!

The final explanation is:
In `getNumAttrs` we declare a variable `h5info` on the stack of type
`H5O_info_t`.
Due to `time_t` being interpreted as 4 bytes, the full `H5O_info_t`
type had a size of 144 bytes *for our Nim object*. But the HDF5
library expects the object to be 160 bytes. We hand a pointer to the
stack object to the Nim library, which writes to it.
As a result of our object being too small, the HDF5 library caused a
stack corruption leading to very bizarre bugs.

My favorite:

```nim
let isItNil = h5attrs.isNil
doAssert not h5attrs.isNil, "But was it nil?" & $isItNil
```

which failed in the assertion, but printed `isItNil = false`...

The confusing part was that the API changed between 1.10 (is what I
run locally on linux) and 1.12 (I run 1.14 on Windows), causing me to
think the problem is in the difference between the `H5O_info_t`
types. But that's not the actual problem, because the type we actually
map to is `H5O_info1_t`, which is backwards compatible, because we
call the `H5Oget_info2` (and friends) procedure, which takes that type
as an argument. The new `H5O_info2_t` type is used in the
`H5Oget_info3` function.